\name{functions}
\alias{functions}
\title{functions}
\description{function to log console outputs to the log file}
\usage{
functions(x)
logger()
}
\arguments{
  \item{x}{the script input}
  \item{y}{the console output}
}
}
\value{
the function output
}

\author{
Aishwarya Shivakumar
}

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
