\name{logger}
\alias{logger}
\title{logger}
\description{
define a function called logger that takes two arguments:
msg (the message to log) and file_name (the name of the log file).
Inside the function, use the cat function to append the message to the log
file, along with the current date and time
}
\usage{
logger(x)
}
\arguments{
  \item{timestamp}{msg (the message to log) and file_name (the name of the log file)}
  \item{message}{Inside the function, use the cat function to append the message
        to the log file, along with the current date and time}
}
}
\author{
Aishwarya Shivakumar
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
