\name{Onload}
\alias{Onload}
\title{Onload}
\description{
define a function called .onLoad that sets the default log file name}
\usage{
Onload(x)
}
\arguments{
  \item{.onLoad}{.onLoad sets the default log file name}
}
\author{
Aishwarya Shivakumar
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
