\name{Retrieving file}
\alias{Retrieving file}
\title{Retrieving file}
\description{
When you want to access the log file from your package, use the getOption
function to retrieve the log file name from the package options
}
\usage{
Retrieving file(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{log file}{use getOption to retrieve log file}

\author{
Aishwarya Shivakumar
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
